5767 - Repository: Clean up UI and component hiearchy (#5826)
* 5767 - Repository: Clean up UI and component hiearchy

* 5767 - Repository: Move title outside and lighten headers

* 5767 - Repository: Adjust title spacing

* 5767 - Repository: Adjust column labels row color
