+++ src/client/pages/CountryHome/Repository/Repository.scss
@@ -13,6 +13,10 @@
     margin: $spacing-s auto;
     width: 50%;
   }
+
+  h3 {
+    margin-bottom: $spacing-xxs;
+  }
 }
 
 .repository__header {
