--- src/client/pages/CountryHome/Repository/RepositoryList/Filters/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Filters'
