+++ src/client/pages/CountryHome/Repository/RepositoryList/Header/ButtonBack/index.ts
@@ -1,0 +1,1 @@
+export { default } from './ButtonBack'
