+++ src/client/pages/CountryHome/Repository/RepositoryList/RepositoryList.tsx
@@ -7,10 +7,9 @@ import { useFolderNavigation } from './hooks/useFolderNavigation'
 import { useGetItems } from './hooks/useGetItems'
 import { useItems } from './hooks/useItems'
 import { useOnToggle } from './hooks/useOnToggle'
-import Breadcrumb from './Breadcrumb'
+import ColumnHeaders from './ColumnHeaders'
 import { RepositoryListContext } from './context'
-import Filters from './Filters'
-import Toolbar from './Toolbar'
+import Header from './Header'
 
 type Props = {
   isGlobal?: boolean
@@ -27,9 +26,8 @@ const RepositoryList: React.FC<Props> = (props) => {
   return (
     <RepositoryListContext.Provider value={contextValue}>
       <div className="repository-list">
-        <Filters isGlobal={isGlobal} />
-        <Breadcrumb />
-        <Toolbar isGlobal={isGlobal} />
+        <Header isGlobal={isGlobal} />
+        <ColumnHeaders isGlobal={isGlobal} />
         {visibleItems.map((item) => (
           <RepositoryListItem key={item.uuid} item={item} />
         ))}
