+++ src/client/pages/CountryHome/Repository/RepositoryList/RepositoryListItem/Folder/Folder.scss
@@ -24,3 +24,7 @@
     }
   }
 }
+
+.repository-list-item--folder > button {
+  grid-column: 1 / -1;
+}
