--- src/client/pages/CountryHome/Repository/RepositoryList/RepositoryListItem/FolderRow/FolderRow.scss
@@ -1,5 +1,0 @@
-@use 'src/client/style/partials' as *;
-
-.repository-list-item--folder > button {
-  grid-column: 1 / -1;
-}
