--- src/client/pages/CountryHome/Repository/RepositoryList/RepositoryListItem/FolderRow/index.ts
@@ -1,1 +1,0 @@
-export { default } from './FolderRow'
