--- src/client/pages/CountryHome/Repository/RepositoryList/Toolbar/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Toolbar'
